requirements.txt
Flask
TensorFlow
scikit-learn
mysql-connector-python
